Hide print statements from ntlink_filter_sequences.py, fix calls to args
